--- feature/search/src/main/java/com/sumin/feature/search/HospitalListUiState.kt
@@ -1,17 +1,0 @@
-package com.sumin.feature.search
-
-data class HospitalListUiState(
-    val items: List<HospitalItemUiState> = listOf(),
-    val messages: List<Message> = listOf(),
-    val isFetchingHospitals: Boolean = false
-)
-
-data class Message(val id: Long, val message: String)
-
-data class HospitalItemUiState(
-    val id: String,
-    val codeName: String,
-    val hospitalName: String,
-    val sidoAddr: String,
-    val sgguAddr: String
-)
