+++ feature/search/src/main/java/com/sumin/feature/search/SearchFragment.kt
@@ -6,12 +6,14 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.appcompat.widget.SearchView
+import androidx.core.view.isVisible
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
+import androidx.paging.LoadState
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.sumin.feature.search.databinding.FragmentSearchBinding
@@ -32,10 +34,6 @@ class SearchFragment : Fragment(), Navigatable {
     private val binding: FragmentSearchBinding get() = requireNotNull(_binding)
 
     private val hospitalAdapter = HospitalAdapter { id ->
-        /*val request = NavDeepLinkRequest.Builder
-            .fromUri("android-app://com.sumin.hospital_detail/hospital_detail_fragment/?hospitalId=$id".toUri())
-            .build()
-        findNavController().navigate(request)*/
         NavigatorMediator.navigate(Route.ActionSearchFragmentToHospitalDetailFragment(id))
     }
 
@@ -67,6 +65,12 @@ class SearchFragment : Fragment(), Navigatable {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         initView()
+
+        viewLifecycleOwner.lifecycleScope.launch {
+            hospitalAdapter.loadStateFlow.collectLatest { loadStates ->
+                binding.progressBar.isVisible = loadStates.refresh is LoadState.Loading
+            }
+        }
     }
 
     private fun initView() {
